pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "project-summary"
version = "0.1.0"
description = "Tool for generating project structure and content summaries"
readme = "README.md"
authors = [{ name = "Fedorello", email = "fedorstartup@gmail.com" }]
license = { file = "LICENSE" }
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Software Development :: Documentation",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
keywords = ["documentation", "project", "summary", "structure", "analysis"]
dependencies = [
    "pyyaml>=6.0",
    "pathlib>=1.0",
]

[project.urls]
Homepage = "https://github.com/fedorello/project-summary"
Repository = "https://github.com/fedorello/project-summary.git"
Issues = "https://github.com/fedorello/project-summary/issues"
Changelog = "https://github.com/fedorello/project-summary/blob/main/CHANGELOG.md"

[project.scripts]
project-summary = "project_summary.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/project_summary"]

[tool.hatch.version]
path = "src/project_summary/__init__.py"